Refactor: Account creation moved out of of saga...
The whole logic occurs in the React component, which is perhaps better,
as the logic is in one place. React Strict mode removed, as it conflicts
iwth MUI presently. MUI Dialogs have been added to inform about
success/failure of the creation request. All relevant state is local to
the react component.